Move time interpolators into separate modules.
