+++ pymt/framework/bmi.py
@@ -1,0 +1,7 @@
+#! /usr/bin/env python
+
+
+def bmi_call(func, *args):
+    rtn = func(*args)
+
+    return rtn
